Changed the reduction function to a proper reduction and added other small fixes.

+++ example/reduce/src/kernel.hpp
@@ -88,16 +88,16 @@ struct ReduceKernel
             alpaka::block::shared::st::allocVar<cheapArray<T, TBlockSize>,
                                                 __COUNTER__>(acc));
 
-        const uint32_t blockIndex(
-            alpaka::idx::getIdx<alpaka::Grid, alpaka::Blocks>(acc)[0]);
-        const uint32_t threadIndex(
-            alpaka::idx::getIdx<alpaka::Block, alpaka::Threads>(acc)[0]);
-        const uint32_t gridDimension(
-            alpaka::workdiv::getWorkDiv<alpaka::Grid, alpaka::Blocks>(acc)[0]);
+        const uint32_t blockIndex(static_cast<uint32_t>(
+            alpaka::idx::getIdx<alpaka::Grid, alpaka::Blocks>(acc)[0]));
+        const uint32_t threadIndex(static_cast<uint32_t>(
+            alpaka::idx::getIdx<alpaka::Block, alpaka::Threads>(acc)[0]));
+        const uint32_t gridDimension(static_cast<uint32_t>(
+            alpaka::workdiv::getWorkDiv<alpaka::Grid, alpaka::Blocks>(acc)[0]));
 
         // equivalent to blockIndex * TBlockSize + threadIndex
-        const uint32_t linearizedIndex(
-            alpaka::idx::getIdx<alpaka::Grid, alpaka::Threads>(acc)[0]);
+        const uint32_t linearizedIndex(static_cast<uint32_t>(
+            alpaka::idx::getIdx<alpaka::Grid, alpaka::Threads>(acc)[0]));
 
         typename GetIterator<T, TElem, TAcc>::Iterator it(
             acc, source, linearizedIndex, gridDimension * TBlockSize, n);
